realtek-poe: clean up coding style
The old convention in realtek-poe was to leave John's old code alone,
with the return value and function name on separate lines, but use a
sane style for modified or new code. This was supposed to be
temporary, but two years later, it is still the case.

Make teh style consistent, and use a single line for most function
definitions. The onle place it makes sense to space things out is for
functions with very large argument lists.

Lecture time:

C defines the "execution character sets" as 26 uppercase and lowercase
Latin alphabet letters, 10 decimal digits, and 29 graphic characters.
A "char" is intended to store these printable characters.
Now for the fun part 6.2.5.3:

    "  If any other character is stored in a char object, the
       resulting value is implementation-defined ... "

Rant time:

"char" is not meant to store integers, so why on earth do porgrammers
insist on using it when they mean 8-bit unsigned interegr? Just use
uint8_t consistently, and end this char madness.

Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>