realtek-poe: move configuration out of global scope
The poe configuration is used in ubus and uloop callbacks. Much of
the realtek-poe state had to be stored in global variables to be
accessed from these callbacks. However, the use of globals isn't
strictly necessary.

Instead of a global structure, use struct containment to keep relevant
ubus objects and the config in a new "poe_context" structure. Use
container_of() to get back a pointer to this context. This way, there
is always a way to get the context from a ubus callback.

Create the config as part of the poe_ctx in main(), and drop the
global config structure.

Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>