realtek-poe: separate MCU state into its own struct
The "state" structure comprised of two states. The first is the state
of the MCU and power delivery. The second is the state of
communication with the MCU, specifically the timeout counters. These
are distinct objects, which are handled by different parts of the code.

To show this difference, move the MCU state to its own data structure.
remove global variables form the reply handlers, and take in the MCU
state as a parameter. Use a const struct pointer in ubus_poe_info_cb
to show that the state is not modified.

One advantage of this separation is that the mcu_state can be moved to
the common header without needing to depend on ubus.h. It is also a
prerequisite for separating the MCU dialect code from the comms code.

To prevent confusion between the mcu state, and mcu representation,
rename the "state" variable to "main_mcu"

Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>