realtek-poe: move main_mcu out of global context
Now that the poe context can be accessed from the ubus context, there
is no longer any need to have global variables. To take advantage of
that, move the mcu data structure to the poe context.

As as a result of movinf `main_mcu`, poe_cmd_queue() can no longer be
used. Thus command functions now need to take the mcu as a paramater,
resulting in a somewhat large change.

Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>